+++ src/app/(sub)/calendar/page.tsx
@@ -1,4 +1,4 @@
-import DatePicker from "@/components/modal/calendar/DatePicker";
+import DatePicker from "@/app/(sub)/calendar/DatePicker";
 import LikeModalView from "../LikeModalView";
 
 export default function DatePickPage() {
