+++ src/app/(sub)/mealplan/add/division/page.tsx
@@ -1,5 +1,5 @@
 import LikeModalView from "@/app/(sub)/LikeModalView";
-import DivisionSelectComponent from "@/components/modal/division/DivistionSelectComponent";
+import DivisionSelectComponent from "./DivistionSelectComponent";
 
 export default function CategorySelectPage() {
   return (
