+++ src/app/(sub)/mealplan/add/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import MealplanForm from "@/components/main/MealplanForm";
+import MealplanForm from "@/app/(sub)/mealplan/MealplanForm";
 
 export default function MealAddPage() {
   const [selectedType, setSelectedType] = useState(0);
