+++ src/app/(sub)/notice/edit/page.tsx
@@ -1,4 +1,4 @@
-import NoticeAddForm from "@/components/modal/notice/NoticeAddForm";
+import NoticeAddForm from "@/app/(sub)/notice/NoticeAddForm";
 import LikeModalView from "../../LikeModalView";
 
 export default function NoticeEditPage() {
