+++ src/app/@modal/(.)notice/edit/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import Modal from "@/components/modal/Modal";
-import NoticeAddForm from "@/components/modal/notice/NoticeAddForm";
+import Modal from "@/app/@modal/Modal";
+import NoticeAddForm from "@/app/(sub)/notice/NoticeAddForm";
 
 export default function NoticeEditModal() {
   return (
